requirements.txt
requests
selenium
schedule
python-dotenv
beautifulsoup4
azure-communication-email >=1.0.0, <1.1.0
azure-communication-sms==1.0.1